feat: adiciona tag picture ao authformscreen
