+++ src/components/AuthFormScreen/index.tsx
@@ -5,7 +5,10 @@ export const AuthFormScreen = () => {
   return (
     <ContentWrapper>
       <section className="logo">
-        <img src="/logo-01.svg" alt="Imagem da logo" />
+        <picture>
+          <source srcSet="/logo-01.svg" media="(min-width: 1024px)" />
+          <img src="/logo-02.png" alt="Imagem da logo" />
+        </picture>
       </section>
 
       <section>
