+++ src/const/sizes.ts
@@ -1,0 +1,11 @@
+export const sizeToPadding = {
+  small: '0.8rem 1.6rem',
+  medium: '0.8rem 3.2rem',
+  large: '1.6rem 0',
+}
+
+export const sizeToWidth = {
+  small: 'auto',
+  medium: '19.5rem',
+  large: '48rem',
+}
