+++ src/types/index.ts
@@ -1,3 +1,5 @@
+import { ButtonHTMLAttributes, ReactNode } from 'react'
+
 export interface ProductsProps {
   _id: string
   amount: number
@@ -11,4 +13,20 @@ export interface ProductsProps {
 export interface CardProps {
   cardData: ProductsProps[]
   onBuyClick: (productId: ProductsProps) => void
+}
+
+export interface ButtonCategorieProps
+  extends ButtonHTMLAttributes<HTMLButtonElement> {
+  size: 'small' | 'medium' | 'large'
+  color?: string
+  loading?: boolean
+  children?: ReactNode
+}
+
+export interface ButtonContainerProps {
+  size: 'small' | 'medium' | 'large'
+}
+
+export interface InputContainerProps {
+  size: 'medium' | 'large'
 }
